switched to error/2 from error/1 in exported funcs
Passing a list of arguments in the second parameter, makes them
show up in the stack trace in place of the function arity number.

For referentially transparent functions (which foo_to_bar tend to
be) failures can then be reproduced from just the stack trace!

If using error/1, you must rely on some higher layer to add enough
info to reproduce the crash.