jsx_to_term:parse_config/2: Honor `return_maps` option value
Previously, no matter what the caller specified in options,
`return_maps` was always set to `true`. The only solution was to not
specify `return_maps` at all.

Unfortunately, if the caller is using a wrapper which sets default
options, including `return_maps`, it's impossible to override it. This
prevented the caller from requesting this mode if he was interested in
preserving the order of keys.

While here, add a testcase for this.